Move cell components to 'props'
